Fixing styling issues
