+++ client/app/styles/_commons.scss
@@ -507,6 +507,7 @@ svg title {
 .cf-modal-controls {
   position: relative;
   bottom: 0;
+  padding-bottom: 3rem;
   margin-top: rem(30px);
 }
